Add RSS feed/blog updates
This follows the same pattern as the GitHub contributions (in addition,
extracting the object used to build up) so that the interface is the
same. It's close enough, anyway.

This also refactors the `Readme` object to accept a "key" for the
section -- from here we're able to define any section we like to update.