Add icons to GitHub contributions
These come from the Readme-Icons project, providing SVGs for the icons
used on GitHub that render at the correct size.

To do this, we split out `FeedItem` from `Contribution` with a similar
enough interface to render them later on in the same way -- the object
itself now handles the rendering of the line to allow us to vary the
output accordingly.

Due to the way the GitHub API handles open/merged/draft PRs, the code to
handle it ends up being quite complex. This seemed an adequate solution to
the problem though.

https://readme-workflows.github.io/Readme-Icons/icons/octicons/